For direct visits to tag page.
Redirect 'em to the blog!

Bonus: Added in a fallback for our atypical WordPress blog setup (we are using a category named Blog instead of a blog page which is set as the Posts page in the admin). If someone (maybe me?) fixes that in the future, this is now future-proof. Tested with typical WP setup and our current.